requirements.txt
vertexai~=1.71.0
preferredsoundplayer~=0.0.5
google-cloud-texttospeech~=2.20.0